pre config variáveis de ambiente
