+++ Main/Composers/SignUpComposer.swift
@@ -9,9 +9,18 @@
 import Foundation
 import Domain
 import iOSUi
+import Presentation
+import Validation
+import Domain
 
 public final class SignUpComposer {
-   public static func composeControllerWith(addAccount: AddAccount) -> SignUpViewController {
-        return ControllerFactory.makeSignUp(addAccount: addAccount)
+    public static func composeControllerWith(addAccount: AddAccount) -> SignUpViewController {
+        let controller = SignUpViewController.instantiate()
+        let emailValidatorAdapter = EmailValidatorAdapter()
+        
+        let presenter = SignUpPresenter(alertView: WeakVarProxy(controller), emailValidator: emailValidatorAdapter, addAccount: addAccount, loadingView: WeakVarProxy(controller))
+        
+        controller.signUp = presenter.signUp
+        return controller
     }
 }
