--- Main/Factories/ControllerFactory.swift
@@ -1,27 +1,0 @@
-//
-//  SignUpFactory.swift
-//  Main
-//
-//  Created by Israel Ermel on 05/09/20.
-//  Copyright © 2020 Israel Ermel. All rights reserved.
-//
-
-import Foundation
-import iOSUi
-import Presentation
-import Validation
-import Domain
-
-class ControllerFactory {
-    static func makeSignUp(addAccount : AddAccount) -> SignUpViewController {
-        let controller = SignUpViewController.instantiate()
-        let emailValidatorAdapter = EmailValidatorAdapter()
-                
-        let presenter = SignUpPresenter(alertView: WeakVarProxy(controller), emailValidator: emailValidatorAdapter, addAccount: addAccount, loadingView: WeakVarProxy(controller))
-        
-        controller.signUp = presenter.signUp
-        return controller
-    }
-    
-}
-
