+++ Main/Factories/UseCaseFactory.swift
@@ -12,9 +12,15 @@ import Infra
 import Domain
 
 final class UseCaseFactory {
+    
+    private static let httpClient = AlamofireAdapter()
+    private static let apiBaseUrl = "https://clean-node-api.herokuapp.com/api"
+    
+    private static func makeUrl(path: String) -> URL {
+        return URL(string: "\(apiBaseUrl)/\(path)")!
+    }
+    
     static func makeRemoteAddAccount() -> AddAccount {
-        let alamofireAdapter = AlamofireAdapter()
-        let url = URL(string: "https://clean-node-api.herokuapp.com/api/signup")!
-        return RemoteAddAccount(url: url, httpClient: alamofireAdapter)
+        return RemoteAddAccount(url: makeUrl(path: "signUp"), httpClient: httpClient)
     }
 }
